# candidate-match-export/candidate-match/openapi.yaml
openapi: 3.0.3
info:
  title: Cloudberries Candidate Match API
  version: 1.1.0
  description: |
    API for candidate matching, CV retrieval, embeddings, and health checks.
    Includes normalized health responses and skill-based matching endpoint.
    Generated from controller sources as of this commit.
  contact:
    name: Cloudberries
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /consultants/{userId}:
    get:
      tags: [ Consultants ]
      summary: Get consultant summary by userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Consultant summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultantSummaryDto'
        '404':
          description: Not found
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /matches:
    post:
      tags: [ Matches ]
      summary: Find candidate matches for a project request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchApiRequest'
      responses:
        '200':
          description: List of candidate matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateMatchResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /matches/requests:
    get:
      tags: [ Matches ]
      summary: List project requests (uploaded customer documents) with coverage info
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - in: query
          name: sort
          schema: { type: string, default: "uploadedAt,desc" }
      responses:
        '200':
          description: Paged list of requests with coverage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMatchesListDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /matches/requests/{id}/top-consultants:
    get:
      tags: [ Matches ]
      summary: Get top consultants for a request (AI-enriched)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 5, minimum: 1, maximum: 10 }
      responses:
        '200':
          description: Top consultants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchConsultantDto'
        '404':
          description: Project request not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /matches/requests/{id}/re-analyze:
    post:
      tags: [ Matches ]
      summary: Re-analyze consultants using Gemini 3.0 batch evaluation
      description: |
        Re-analyzes and re-ranks consultants for a project request using improved AI matching.
        This endpoint:
        - Fetches ~30 consultants matching required skills
        - Scores by 50% skills + 50% CV quality
        - Selects top 10 candidates
        - Sends all 10 CVs in ONE Gemini API call for batch evaluation
        - Returns ranked results with detailed justifications
        
        Note: This endpoint runs synchronously and may take up to 60 seconds.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Re-analyzed and ranked consultants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchConsultantDto'
        '404':
          description: Project request not found
        '500':
          description: AI service error
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    CoverageStatus:
      type: string
      enum: [ GREEN, YELLOW, RED, NEUTRAL ]
    MatchesListItemDto:
      type: object
      required: [ id, hitCount, coverageStatus, coverageLabel ]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          nullable: true
        customerName:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
          nullable: true
        deadlineDate:
          type: string
          format: date-time
          nullable: true
        hitCount:
          type: integer
          format: int64
        coverageStatus:
          $ref: '#/components/schemas/CoverageStatus'
        coverageLabel:
          type: string
    PagedMatchesListDto:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/MatchesListItemDto' }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        currentPage: { type: integer }
        pageSize: { type: integer }
        hasNext: { type: boolean }
        hasPrevious: { type: boolean }
    MatchConsultantDto:
      type: object
      required: [ userId, name, cvId, relevanceScore ]
      properties:
        userId: { type: string }
        name: { type: string }
        cvId: { type: string }
        relevanceScore:
          type: number
          format: double
        justification:
          type: string
          nullable: true
...
    ConsultantSummaryDto:
      type: object
      required: [ userId, name, email, bornYear, defaultCvId ]
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        bornYear:
          type: integer
        defaultCvId:
          type: string
...
    MatchApiRequest:
      type: object
      required: [ projectRequestText ]
      properties:
        projectRequestText:
          type: string
          description: Project request text/requirements
...
    CandidateMatchResponse:
      type: object
      required: [ totalScore, summary ]
      properties:
        totalScore:
          type: string
          description: Overall match score (string in current implementation)
        summary:
          type: string
        matchTimeSeconds:
          type: integer
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/Requirement'
    Requirement:
      type: object
      required: [ name, comment, score ]
      properties:
        name:
          type: string
        comment:
          type: string
        score:
          type: string
...
    Problem:
      type: object
      description: Generic error payload
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # AI Chat Search Schemas
...